Don't require jinja for migration doctests
Moving the jinja2 import out of migration_utils.py allows running
doctest_migration.py in the CI without having to install doc
dependencies.
